requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.5.0

# French NLP models
sentence-transformers>=2.2.0
transformers>=4.30.0

# Dashboard
streamlit>=1.25.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
reportlab>=4.0.0
tqdm>=4.65.0

# Text analysis
textstat>=0.7.3
word2vec>=0.11.1
gensim>=4.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0
multiprocessing-logging>=0.3.4

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0